+++ scout/server/blueprints/panels/views.py
@@ -3,7 +3,16 @@ import datetime
 import json
 import logging
 
-from flask import Blueprint, Response, flash, redirect, render_template, request, send_file, url_for
+from flask import (
+    Blueprint,
+    Response,
+    flash,
+    redirect,
+    render_template,
+    request,
+    send_file,
+    url_for,
+)
 from flask_login import current_user
 from markupsafe import escape
 
@@ -38,7 +47,9 @@ def api_panels(panel_name):
     """
     json_out = controllers.get_panels(store, panel_name)
 
-    return Response(json.dumps(json_out, default=jsonconverter), mimetype="application/json")
+    return Response(
+        json.dumps(json_out, default=jsonconverter), mimetype="application/json"
+    )
 
 
 @panels_bp.route("/panels", methods=["GET", "POST"])
@@ -58,7 +69,8 @@ def panels():
             hgnc_id = hgnc_symbols.pop()
         except ValueError:
             flash(
-                "Provided gene info could not be parsed! " "Please allow autocompletion to finish.",
+                "Provided gene info could not be parsed! "
+                "Please allow autocompletion to finish.",
                 "warning",
             )
         panels_found = store.search_panels_hgnc_id(hgnc_id)
@@ -76,9 +88,9 @@ def panels():
     panel_names = [
         name
         for institute in institutes
-        for name in store.gene_panels(institute_id=institute["_id"], include_hidden=True).distinct(
-            "panel_name"
-        )
+        for name in store.gene_panels(
+            institute_id=institute["_id"], include_hidden=True
+        ).distinct("panel_name")
     ]
     panel_versions = {}
     for name in panel_names:
@@ -90,7 +102,9 @@ def panels():
         institute_panels = []
         for panel in store.latest_panels(institute_obj["_id"], include_hidden=True):
             panel["writable"] = (
-                "" if controllers.panel_write_granted(panel, current_user) else "disabled"
+                ""
+                if controllers.panel_write_granted(panel, current_user)
+                else "disabled"
             )
             institute_panels.append(panel)
 
@@ -147,10 +161,14 @@ def panel(panel_id):
         if action == "add":
             panel_gene = controllers.existing_gene(store, panel_obj, hgnc_id)
             if panel_gene:
-                flash("gene already in panel: {}".format(panel_gene["symbol"]), "warning")
+                flash(
+                    "gene already in panel: {}".format(panel_gene["symbol"]), "warning"
+                )
             else:
                 # ask user to fill-in more information about the gene
-                return redirect(url_for(".gene_edit", panel_id=panel_id, hgnc_id=hgnc_id))
+                return redirect(
+                    url_for(".gene_edit", panel_id=panel_id, hgnc_id=hgnc_id)
+                )
         elif action == "delete":
             LOG.debug("marking gene to be deleted: %s", hgnc_id)
             panel_obj = store.add_pending(panel_obj, gene_obj, action="delete")
@@ -165,7 +183,9 @@ def panel(panel_id):
         panel_name = panel_obj["panel_name"]
         latest_panel = store.gene_panel(panel_name)
         if panel_obj["version"] < latest_panel["version"]:
-            extra_genes, missing_genes = check_outdated_gene_panel(panel_obj, latest_panel)
+            extra_genes, missing_genes = check_outdated_gene_panel(
+                panel_obj, latest_panel
+            )
             if extra_genes or missing_genes:
                 case_obj["outdated_panels"][panel_name] = {
                     "missing_genes": missing_genes,
@@ -365,7 +385,9 @@ def tx_choices(hgnc_id, panel_obj):
         if gene_obj:
             for transcript in gene_obj.get("disease_associated_transcripts", []):
                 if (transcript, transcript) not in transcript_choices:
-                    transcript_choices.append((transcript, f"{transcript} (previous choice)"))
+                    transcript_choices.append(
+                        (transcript, f"{transcript} (previous choice)")
+                    )
     return transcript_choices
 
 
@@ -384,19 +406,21 @@ def gene_edit(panel_id, hgnc_id):
     form.disease_associated_transcripts.choices = tx_choices(hgnc_id, panel_obj)
     if form.validate_on_submit():
         action = "edit" if panel_gene else "add"
-        info_data = form.data.copy()
-        if "csrf_token" in info_data:
-            del info_data["csrf_token"]
-        if info_data["custom_inheritance_models"] != "":
-            info_data["custom_inheritance_models"] = info_data["custom_inheritance_models"].split(
-                ","
+        info_data: dict = {k: v for k, v in form.data.copy().items() if v}
+        info_data.pop("csrf_token", None)
+
+        if info_data.get("custom_inheritance_models", ""):
+            info_data["custom_inheritance_models"] = (
+                info_data["custom_inheritance_models"].replace(" ", "").split(",")
             )
         store.add_pending(panel_obj, hgnc_gene, action=action, info=info_data)
         return redirect(url_for(".panel", panel_id=panel_id))
 
     if panel_gene:
         custom_models = [
-            model for model in panel_gene.get("custom_inheritance_models", []) if model != ""
+            model
+            for model in panel_gene.get("custom_inheritance_models", [])
+            if model != ""
         ]
         if custom_models:
             form.custom_inheritance_models.data = ", ".join(
@@ -416,4 +440,5 @@ def gene_edit(panel_id, hgnc_id):
                 panel_value = panel_gene.get(field_key)
                 if panel_value is not None:
                     form_field.process_data(panel_value)
+
     return dict(panel=panel_obj, form=form, gene=hgnc_gene, panel_gene=panel_gene)